chore: move custom_progress_indicator.dart under widgets
